All Beatjazz prosthetic patches
this is the entire list of patches that went into the beatjazz
prosthesis.  the primary ones are:

Beatjazzviz1
Bjsynthloopermod1
Bjcontrolmod

all others are pulled into these three main patches.

currently, those 3 patches must be opened in the order stated above,
each in its own instance of PD-extended.  this is so that each patch can
be assigned its own core (if you have a multi-core computer, which, to
run this patch, you should)